+++ src/main/java/com/ats/tril/model/GetPODetail.java
@@ -30,19 +30,19 @@ public class GetPODetail {
 	private String itemUom;
 	
 	@Column(name = "item_qty")
-	private int itemQty;
+	private float itemQty;
 	
 	@Column(name = "item_rate")
 	private float itemRate;
 	
 	@Column(name = "mrn_qty")
-	private int mrnQty;
+	private float mrnQty;
 	
 	@Column(name = "pending_qty")
-	private int pendingQty;
+	private float pendingQty;
 	
 	@Column(name = "inded_qty")
-	private int indedQty;
+	private float indedQty;
 	
 	@Column(name = "disc_per")
 	private float discPer;
@@ -111,19 +111,19 @@ public class GetPODetail {
 	public String getItemUom() {
 		return itemUom;
 	}
-	public int getItemQty() {
+	public float getItemQty() {
 		return itemQty;
 	}
 	public float getItemRate() {
 		return itemRate;
 	}
-	public int getMrnQty() {
+	public float getMrnQty() {
 		return mrnQty;
 	}
-	public int getPendingQty() {
+	public float getPendingQty() {
 		return pendingQty;
 	}
-	public int getIndedQty() {
+	public float getIndedQty() {
 		return indedQty;
 	}
 	public float getDiscPer() {
@@ -201,19 +201,19 @@ public class GetPODetail {
 	public void setItemUom(String itemUom) {
 		this.itemUom = itemUom;
 	}
-	public void setItemQty(int itemQty) {
+	public void setItemQty(float itemQty) {
 		this.itemQty = itemQty;
 	}
 	public void setItemRate(float itemRate) {
 		this.itemRate = itemRate;
 	}
-	public void setMrnQty(int mrnQty) {
+	public void setMrnQty(float mrnQty) {
 		this.mrnQty = mrnQty;
 	}
-	public void setPendingQty(int pendingQty) {
+	public void setPendingQty(float pendingQty) {
 		this.pendingQty = pendingQty;
 	}
-	public void setIndedQty(int indedQty) {
+	public void setIndedQty(float indedQty) {
 		this.indedQty = indedQty;
 	}
 	public void setDiscPer(float discPer) {
