+++ src/main/java/com/ats/tril/model/indent/GetIndentDetail.java
@@ -36,7 +36,7 @@ public class GetIndentDetail {
 	
 	private int itemId;
 	
-	private int indQty;
+	private float indQty;
 	
 	private String indItemUom;
 	
@@ -54,7 +54,7 @@ public class GetIndentDetail {
 	
 	private int delStatus;
 	
-	private int	indFyr;
+	private float	indFyr;
 	
 	private String itemDesc;
 	private String itemCode;
@@ -74,7 +74,7 @@ public class GetIndentDetail {
 	public int getItemId() {
 		return itemId;
 	}
-	public int getIndQty() {
+	public float getIndQty() {
 		return indQty;
 	}
 	public String getIndItemUom() {
@@ -103,7 +103,7 @@ public class GetIndentDetail {
 	public int getDelStatus() {
 		return delStatus;
 	}
-	public int getIndFyr() {
+	public float getIndFyr() {
 		return indFyr;
 	}
 	public String getItemDesc() {
@@ -130,7 +130,7 @@ public class GetIndentDetail {
 	public void setItemId(int itemId) {
 		this.itemId = itemId;
 	}
-	public void setIndQty(int indQty) {
+	public void setIndQty(float indQty) {
 		this.indQty = indQty;
 	}
 	public void setIndItemUom(String indItemUom) {
@@ -157,7 +157,7 @@ public class GetIndentDetail {
 	public void setDelStatus(int delStatus) {
 		this.delStatus = delStatus;
 	}
-	public void setIndFyr(int indFyr) {
+	public void setIndFyr(float indFyr) {
 		this.indFyr = indFyr;
 	}
 	public void setItemDesc(String itemDesc) {
