+++ src/main/java/com/ats/tril/model/mrn/GetMrnDetailRej.java
@@ -22,23 +22,23 @@ public class GetMrnDetailRej {
 
 	private String itemCode;
 
-	private int indentQty;
+	private float indentQty;
 
-	private int poQty;
+	private float poQty;
 
-	private int mrnQty;
+	private float mrnQty;
 
-	private int approveQty;
+	private float approveQty;
 
-	private int rejectQty;
+	private float rejectQty;
 
 	private int rejectRemark;
 
 	private String batchNo;
 
-	private int issueQty;
+	private float issueQty;
 
-	private int remainingQty;
+	private float remainingQty;
 
 	private int poId;
 
@@ -88,43 +88,43 @@ public class GetMrnDetailRej {
 		this.itemCode = itemCode;
 	}
 
-	public int getIndentQty() {
+	public float getIndentQty() {
 		return indentQty;
 	}
 
-	public void setIndentQty(int indentQty) {
+	public void setIndentQty(float indentQty) {
 		this.indentQty = indentQty;
 	}
 
-	public int getPoQty() {
+	public float getPoQty() {
 		return poQty;
 	}
 
-	public void setPoQty(int poQty) {
+	public void setPoQty(float poQty) {
 		this.poQty = poQty;
 	}
 
-	public int getMrnQty() {
+	public float getMrnQty() {
 		return mrnQty;
 	}
 
-	public void setMrnQty(int mrnQty) {
+	public void setMrnQty(float mrnQty) {
 		this.mrnQty = mrnQty;
 	}
 
-	public int getApproveQty() {
+	public float getApproveQty() {
 		return approveQty;
 	}
 
-	public void setApproveQty(int approveQty) {
+	public void setApproveQty(float approveQty) {
 		this.approveQty = approveQty;
 	}
 
-	public int getRejectQty() {
+	public float getRejectQty() {
 		return rejectQty;
 	}
 
-	public void setRejectQty(int rejectQty) {
+	public void setRejectQty(float rejectQty) {
 		this.rejectQty = rejectQty;
 	}
 
@@ -144,19 +144,19 @@ public class GetMrnDetailRej {
 		this.batchNo = batchNo;
 	}
 
-	public int getIssueQty() {
+	public float getIssueQty() {
 		return issueQty;
 	}
 
-	public void setIssueQty(int issueQty) {
+	public void setIssueQty(float issueQty) {
 		this.issueQty = issueQty;
 	}
 
-	public int getRemainingQty() {
+	public float getRemainingQty() {
 		return remainingQty;
 	}
 
-	public void setRemainingQty(int remainingQty) {
+	public void setRemainingQty(float remainingQty) {
 		this.remainingQty = remainingQty;
 	}
